# a few perf improvements