--- a/editor.ts
+++ b/editor.ts
@@ -923,6 +923,7 @@ namespace microcode {
         }
 
         /* override */ draw() {
+            control.enablePerfCounter("rule.draw")
             // don't render when out of screenR
             if (
                 this.xfrm.worldPos.y + this.bounds.top > Screen.BOTTOM_EDGE ||
@@ -931,9 +932,7 @@ namespace microcode {
                 return
             }
 
-            Screen.fillBoundsXfrm(this.xfrm, this.bounds, 11)
-            Screen.fillBoundsXfrm(this.xfrm, this.whenBounds, 13)
-            Screen.outlineBoundsXfrm(this.xfrm, this.bounds, 1, 12)
+            this.drawBackground()
             this.handleBtn.draw()
             if (this.whenInsertBtn) this.whenInsertBtn.draw()
             this.arrow.draw()
@@ -942,6 +941,13 @@ namespace microcode {
                 const buttons = this.rule[name]
                 buttons.forEach(btn => btn.draw())
             })
+        }
+
+        private drawBackground() {
+            control.enablePerfCounter("rule.draw.bkg")
+            Screen.fillBoundsXfrm(this.xfrm, this.bounds, 11)
+            Screen.fillBoundsXfrm(this.xfrm, this.whenBounds, 13)
+            Screen.outlineBoundsXfrm(this.xfrm, this.bounds, 1, 12)
         }
     }
 }