--- a/options.ts
+++ b/options.ts
@@ -2,6 +2,6 @@ namespace microcode {
     export class Options {
         public static toolTips: boolean = true
         public static repeatKey: boolean = false
-        public static fps: boolean = false
+        public static fps: boolean = true
     }
 }