Baseline 1.2.3: rock paper scissors, 9.5fps, 5.4fps with led editor
- drop shadow on tiles: 2fps
- drop bounds checking on all sprites: 2fps