@pelikhan 2022-09-28
rps: 14.6fps
leds: 7.2fps