@pelikhan 2022-09-28
rps: 15.9fps
leds: 8.9fps