@pelikhan 2022-09-28
rps: 16.7fps
leds: 8fps